+++ src/service/api/user/unfollow.post.ts
@@ -14,6 +14,8 @@ const POST_Unfollow = async (req: Request, res: Response) => {
     if (hasErrors(req, res)) return;
     const sourceId = res.locals.oauth.token.user._id;
     const targetId = req.body.target;
+    if (sourceId === targetId)
+      return res.status(statusCodes.clientInputError).json(new ErrorResponse(errorMessages.clientInputError));
     const result = await FollowModel.deleteOne({
       sourceId,
       targetId,
